Improve Acceptance tests
